Update outdated comment about margin.
It is no longer a 0.5i margin everywhere.

Yes, I never thought I would make a PR for someone else's resume, but it turns out this is the best looking groff resume in the entire internet, in my humble opinion. So here I am helping fellow groffers to get familiar with the ms internals.