Fix formatting issue with empty item
If an item doesn't contain an list entries, then the right-aligned text
won't be displayed on the same level as the left-aligned text.

This is a basic fix to realign then by manually shifting the
right-aligned text vertically by a fixed amount.

I tried to use .mk and .rt but it simply didn't work.